# openapi/controller-v1.yaml
...
openapi: 3.0.3
info:
  title: Apache TacticalMesh Controller API
  description: |
    REST API for the Apache TacticalMesh Mesh Controller.
    
    This API enables management of mesh nodes, command dispatch, configuration,
    and user authentication for tactical edge networking deployments.
    
    ## Authentication
    
    Most endpoints require JWT Bearer token authentication. Obtain a token via
    the `/api/v1/auth/login` endpoint.
    
    ## Roles
    
    - **admin**: Full access to all operations
    - **operator**: Can manage nodes and send commands
    - **observer**: Read-only access to node status
    
    ## License
    
    Apache License 2.0 - https://www.apache.org/licenses/LICENSE-2.0
  version: 0.1.0
  license:
    name: Apache 2.0
    url: https://www.apache.org/licenses/LICENSE-2.0
  contact:
    name: Apache TacticalMesh
    url: https://github.com/TamTunnel/Apache-TacticalMesh
servers:
  - url: http://localhost:8000
    description: Local development server
  - url: https://api.tacticalmesh.example.com
    description: Production server
...
paths:
...
  /api/v1/auth/login:
    post:
      tags:
        - Authentication
      summary: User login
      description: Authenticate user and obtain JWT access token
      operationId: login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
      responses:
        '200':
          description: Login successful
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Token'
        '401':
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /api/v1/auth/register:
    post:
      tags:
        - Authentication
      summary: Register new user
      description: Create a new user account (admin only)
      operationId: registerUser
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserCreate'
      responses:
        '201':
          description: User created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '400':
          description: Username or email already exists
        '403':
          description: Admin role required
...
  /api/v1/commands:
    get:
      tags:
        - Commands
      summary: List commands
      description: List commands with pagination and filtering
      operationId: listCommands
      security:
        - bearerAuth: []
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: page_size
          in: query
          schema:
            type: integer
            default: 50
        - name: status_filter
          in: query
          schema:
            $ref: '#/components/schemas/CommandStatus'
        - name: command_type
          in: query
          schema:
            $ref: '#/components/schemas/CommandType'
        - name: target_node_id
          in: query
          schema:
            type: string
      responses:
        '200':
          description: List of commands
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CommandListResponse'
    post:
      tags:
        - Commands
      summary: Create command
      description: Create a new command for a target node
      operationId: createCommand
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CommandCreate'
      responses:
        '201':
          description: Command created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Command'
        '404':
          description: Target node not found
  /api/v1/commands/{command_id}:
    get:
      tags:
        - Commands
      summary: Get command details
      description: Get details and status of a specific command
      operationId: getCommand
      security:
        - bearerAuth: []
      parameters:
        - name: command_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Command details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Command'
        '404':
          description: Command not found
    delete:
      tags:
        - Commands
      summary: Cancel command
      description: Cancel a pending command
      operationId: cancelCommand
      security:
        - bearerAuth: []
      parameters:
        - name: command_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '204':
          description: Command cancelled
        '400':
          description: Command cannot be cancelled
        '404':
          description: Command not found
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    Error:
      type: object
      properties:
        detail:
          type: string
...
    LoginRequest:
      type: object
      required:
        - username
        - password
      properties:
        username:
          type: string
          minLength: 3
        password:
          type: string
          minLength: 8
    Token:
      type: object
      properties:
        access_token:
          type: string
        token_type:
          type: string
          example: bearer
        expires_in:
          type: integer
        role:
          $ref: '#/components/schemas/UserRole'
    UserRole:
      type: string
      enum: [admin, operator, observer]
    UserCreate:
      type: object
      required:
        - username
        - password
      properties:
        username:
          type: string
        email:
          type: string
          format: email
        password:
          type: string
          minLength: 8
        role:
          $ref: '#/components/schemas/UserRole'
    User:
      type: object
      properties:
        id:
          type: string
          format: uuid
        username:
          type: string
        email:
          type: string
        role:
          $ref: '#/components/schemas/UserRole'
        is_active:
          type: boolean
        created_at:
          type: string
          format: date-time
        last_login:
          type: string
          format: date-time
...
    CommandType:
      type: string
      enum: [ping, reload_config, update_config, change_role, custom]
    CommandStatus:
      type: string
      enum: [pending, sent, acknowledged, executing, completed, failed, timeout]
    CommandCreate:
      type: object
      required:
        - target_node_id
        - command_type
      properties:
        target_node_id:
          type: string
        command_type:
          $ref: '#/components/schemas/CommandType'
        payload:
          type: object
...
    Command:
      type: object
      properties:
        id:
          type: string
          format: uuid
        command_type:
          $ref: '#/components/schemas/CommandType'
        status:
          $ref: '#/components/schemas/CommandStatus'
        target_node_id:
          type: string
          format: uuid
        payload:
          type: object
        result:
          type: object
        error_message:
          type: string
        created_by:
          type: string
          format: uuid
        created_at:
          type: string
          format: date-time
        sent_at:
          type: string
          format: date-time
        acknowledged_at:
          type: string
          format: date-time
        completed_at:
          type: string
          format: date-time
    CommandListResponse:
      type: object
      properties:
        commands:
          type: array
          items:
            $ref: '#/components/schemas/Command'
        total:
          type: integer
        page:
          type: integer
        page_size:
          type: integer
...